Move store-related functions from request to store
